requirements.txt
# Optional: required for gmail_oauth_extract.py (bulk Gmail receipt scanning).
# The core dashboard (analyze.py, server.py, live_monitor.py, snapshot.py)
# uses only the Python standard library — no dependencies needed for basic use.

google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
